add naming and resolution support
Naming is valuable and necessary to support resolution. Use =Name to
specify a name for the current item (not sure what happens if you do it
twice, don't do that); use @Name to reference the name in a resolution.
References can also be used in other places and will eventually be
parsed and dropped in a separate field.

Resolutions allow you to have open and resolved #todo items so you can
see at a glance what #todo items haven't been taken care of yet using
the -Unresolved switch to Find-Daylog. Use :resolves @whatever to
resolve the @whatever todo item, at which point it will no longer
appear, and the =whatever item will get a Resolved field set to true.