reformat Find-Daylog to make logic more clear
Previously options that set other parameters but didn't actually filter
anything were intermixed with actual filters.

It would be possible to get a lot of these converted to some kind
of hash table since they're all doing the same Where-Object thing,
but I'm not sure it's worth it right now.