support "Hats"
These will serve a similar function to my categories in Outlook. For
instance, I might want to give a TFS build issue with a VB app -- the VB
hat and the TFS hat. There will definitely be SS*S, VB, and TFS hats;
maybe a PowerShell one too. They don't have to exactly correspond and
there's no practical limit on the number I can have, unlike in Outlook
where there are only so many colors and it gets ugly assigning them
without going through a huge menu if you havea bunch of them.

Note that if you want to specify multiple hats, they have to go on
multiple lines; this is a general problem with all parsing where you
can't have more than one of the *same* parsed item on each line because
switch won't hit the same case twice even if it can hit multiple cases.
It just is unlikely you'd ever want to do that with : or $, so I haven't
noticed it until now. I'm not sure if there's a clearly better way to
handle this that doesn't majorly complicate the parser.