add autohat directives
I expect there will be cause for more !directives in the future. We
should also have a way of printing out these directives to verify that
we've done them right.

It might be nice to support some character that could be tacked onto the
end of a $billingItem to assert that it has an autohat -- this would
prevent some possible mistakes and make the log easier to read.