allow :attributes to have multiple values
If there are multiple, they're automatically wrapped into an array,
which fits nicely with how PowerShell typically handles arrays.

This was implemented so that we can say an item :resolves
multiple todo items, but it could be handy for other things too,
like the :with attribute I often use on meetings.