the real optimization
Woot! This actually legitimately helps. I think too much pipelining,
while very elegant and great on memory usage, actually tends to slow
PowerShell down because there's more overhead. So even on the first
run of a session, the new version which reads everything in and then
starts passing it down the pipeline is faster. The real gains, though,
come when you run multiple commands in one session in a row without
changing the file; we now cache the parsed file in memory, so we can go
directly to the filtering/computing steps. It's lightning fast for
the subsequent commands.