limit the need to escape dollar signs
I never, ever put billing items in the middle of lines intentionally,
they always go on their own lines at the top of the entry.
Therefore, we don't match billing at all unless there are no characters
besides whitespace on the line before the dollar sign. This should
greatly reduce the number of erroneous-split-billing errors we get
in day-to-day use.