fix bug where zero-time break entries would be added
First of all, the condition for whether to output an entry compared
against a nonexistent variable, so the item was added unconditionally.
Apparently entries with zero time are filtered out by FDT and FDS, which
I'm always using when I'm paying attention to time readings, so I never
noticed this.

Second of all, when you use Write-Output with a null function return, a
null object is written to the output stream, so anytime the condition
was false this started injecting null objects and wreaking havoc. This
is easy enough to fix by removing the Write-Output, which is just for
ease of understanding.